+++ rust/common/mod.rs
@@ -20,6 +20,7 @@
  *
  */
 
+pub mod date_time;
 pub mod error;
 pub mod string;
 pub mod token;
